requirements.txt
torch
torchvision
timm
numpy
albumentations
opencv-python
Pillow
scikit-learn
matplotlib
fvcore
tensorboard